fix: delete list component and move logic to main component, fix logic for escape handling
