--- src/components/lyne-autocomplete-list/lyne-autocomplete-list.events.ts
@@ -1,4 +1,0 @@
-export default {
-  selected: 'lyne-autocomplete-list_selected',
-  setInputFocus: 'lyne-autocomplete-list_set-input-focus'
-};
