--- src/components/lyne-autocomplete-list/styles/lyne-autocomplete-list.shared.scss
@@ -1,21 +1,0 @@
-//*****************************************************************************
-// Important
-//
-// This is the components shared behaviour: shared
-//
-// If in shared mode, globally defined styles are imported globally and shared
-// over all components and therefore are not explicitly imported in here unlike
-// in *.default.scss)
-//
-// 1. The component styles get imported from *.component.scss
-//
-// 2. Mode overrides get applied
-//*****************************************************************************
-
-// Import component styles
-@import "./lyne-autocomplete-list.component.scss";
-
-// Mode overrides/tests
-// :host {
-//   background-color: pink;
-// }
