+++ src/components/lyne-autocomplete/lyne-autocomplete.events.ts
@@ -1,4 +1,3 @@
 export default {
-  keypress: 'lyne-autocomplete_keypress',
   selected: 'lyne-autocomplete_selected'
 };
