+++ src/components/lyne-autocomplete/readme.md
@@ -27,15 +27,14 @@
 ### Depends on
 
 - [lyne-text-input](../lyne-text-input)
-- [lyne-autocomplete-list](../lyne-autocomplete-list)
+- [lyne-autocomplete-item](../lyne-autocomplete-item)
 
 ### Graph
 ```mermaid
 graph TD;
   lyne-autocomplete --> lyne-text-input
-  lyne-autocomplete --> lyne-autocomplete-list
+  lyne-autocomplete --> lyne-autocomplete-item
   lyne-text-input --> lyne-input-error
-  lyne-autocomplete-list --> lyne-autocomplete-item
   style lyne-autocomplete fill:#f9f,stroke:#333,stroke-width:4px
 ```
