+++ src/components/lyne-autocomplete/styles/lyne-autocomplete.component.scss
@@ -22,3 +22,46 @@
 .autocomplete {
   position: relative;
 }
+
+.autocomplete__list {
+  @include listReset;
+  @include shadow-level-5-hard;
+
+  cursor: default;
+  position: absolute;
+  display: none;
+  overflow-y: scroll;
+  overflow-behavior: contain;
+  top: calc(100% - #{pxToRem(var(--input-height))});
+  right: 0;
+  left: 0;
+  z-index: -1;
+
+  border-radius: pxToRem(var(--border-radius-4x));
+  padding-block-start: calc(#{pxToRem(var(--input-height))} + #{pxToRem(var(--spacing-fixed-3x))});
+  padding-block-end: pxToRem(var(--spacing-fixed-3x));
+  padding-inline-start: pxToRem(var(--spacing-responsive-xxxs-zero));
+  padding-inline-end: pxToRem(var(--spacing-responsive-xxxs-zero));
+
+  @include mq($from: medium) {
+    padding-inline-start: pxToRem(var(--spacing-responsive-xxxs-medium));
+    padding-inline-end: pxToRem(var(--spacing-responsive-xxxs-medium));
+  }
+
+  @include ifForcedColors() {
+    background-color: Canvas;
+    border: pxToRem(1) solid Highlight;
+    border-block-start: none;
+  }
+
+}
+
+.autocomplete__list--visible {
+  display: flex;
+  flex-direction: column;
+  gap: pxToRem(2);
+}
+
+.autocomplete__accessibility-hint {
+  @include screenReaderOnly;
+}
